feat: Add PaymentSuccess and PaymentFailure components
The PaymentSuccess component is added to display a success message after a successful payment transaction. It includes the transaction ID if available and a link to the user dashboard.

The PaymentFailure component is added to display an error message when there is an issue processing the payment. It includes the error message and provides options to try again or go to the user dashboard.